Add a component test for <Main>, rename hooks to snakeCase
The helpers/ use snakeCase,
so that probably makes sense.
