+++ js/src/edit-block/hooks/index.js
@@ -1,4 +1,4 @@
-export { default as useBlock } from './use-block';
-export { default as useField } from './use-field';
-export { default as useCategories } from './use-categories';
-export { default as usePostTypes } from './use-post-types';
+export { default as useBlock } from './useBlock';
+export { default as useField } from './useField';
+export { default as useCategories } from './useCategories';
+export { default as usePostTypes } from './usePostTypes';
